+++ gecode/set/int/card.hpp
@@ -78,10 +78,10 @@ namespace Gecode { namespace Set { namespace Int {
    do {
      x1min = x1.min();
      x1max = x1.max();
-     GECODE_ME_CHECK(x0.cardMin(home,x1min));
-     GECODE_ME_CHECK(x0.cardMax(home,x1max));
-     GECODE_ME_CHECK(x1.gq(home,(int)x0.cardMin()));
-     GECODE_ME_CHECK(x1.lq(home,(int)x0.cardMax()));
+     GECODE_ME_CHECK(x0.cardMin(home,static_cast<unsigned int>(x1min)));
+     GECODE_ME_CHECK(x0.cardMax(home,static_cast<unsigned int>(x1max)));
+     GECODE_ME_CHECK(x1.gq(home,static_cast<int>(x0.cardMin())));
+     GECODE_ME_CHECK(x1.lq(home,static_cast<int>(x0.cardMax())));
    } while (x1.min() > x1min || x1.max() < x1max);
    if (x1.assigned())
      return ES_SUBSUMED(*this,home);
